requirements.txt
Flask==3.1.0
Flask_SocketIO==5.5.1
httpx==0.28.1
kaggle==1.6.17
matplotlib==3.10.0
numpy==2.2.3
opencv_contrib_python_headless==4.10.0.84
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
Pillow==9.5.0
Pillow==11.1.0
playwright==1.49.1
python_engineio==4.11.2
torch==2.5.1+cu124
torchvision==0.20.1
watchdog==6.0.0
